move: 패키지 이동 model -> util
