+++ bank/src/main/java/com/fisa/bank/loan/application/service/LoanService.java
@@ -26,8 +26,8 @@ import com.fisa.bank.loan.application.dto.response.LoanProductCreateResponse;
 import com.fisa.bank.loan.application.dto.response.LoanProductResponse;
 import com.fisa.bank.loan.application.dto.response.PagedResponse;
 import com.fisa.bank.loan.application.exception.*;
-import com.fisa.bank.loan.application.model.EarlyRepayInterestRate;
 import com.fisa.bank.loan.application.model.MonthlyRepayment;
+import com.fisa.bank.loan.application.util.EarlyRepayInterestRate;
 import com.fisa.bank.loan.persistence.entity.*;
 import com.fisa.bank.loan.persistence.entity.id.LoanLedgerId;
 import com.fisa.bank.loan.persistence.entity.id.LoanProductId;
